TMVA: removed warnings in MethodDL

+++ tmva/tmva/src/DNN/Architectures/Reference/DenoisePropagation.cxx
@@ -61,11 +61,9 @@ void TReference<Real_t>::UpdateParams(
    }
 
   //updating fHBiases
-   for(size_t i = 0; i < fHBiases.GetNrows(); i++)
-   {
+   for (size_t i = 0; i < (size_t)fHBiases.GetNrows(); i++) {
       HBiasError(i,0) = 0;
-      for(size_t j = 0; j < fVBiases.GetNrows(); j++)
-      {
+      for (size_t j = 0; j < (size_t)fVBiases.GetNrows(); j++) {
          HBiasError(i, 0) += fWeights(i, j) * VBiasError(j, 0);
       }
       HBiasError(i, 0) *= y(i, 0) * (1 - y(i, 0));
@@ -73,10 +71,8 @@ void TReference<Real_t>::UpdateParams(
    }
 
    //updating weights
-   for(size_t i = 0; i < fHBiases.GetNrows(); i++)
-   {
-      for(size_t j = 0; j< fVBiases.GetNrows(); j++)
-      {
+   for (size_t i = 0; i < (size_t)fHBiases.GetNrows(); i++) {
+      for (size_t j = 0; j < (size_t)fVBiases.GetNrows(); j++) {
          fWeights(i, j) += learningRate * (HBiasError(i, 0) * tildeX(j, 0) +
                                           VBiasError(j, 0) * y(i, 0)) / fBatchSize;
       }
